# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/registration": {
            "post": {
                "operationId": "ProfilesController_registration",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserProfileDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "put": {
                "operationId": "ProfilesController_updateProfile",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserProfileDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "defaultBearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "ProfilesController_deleteProfile",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "defaultBearerAuth": []
                    }
                ]
            }
        },
...
        "/genre": {
            "post": {
                "operationId": "GenreController_createGenre",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGenreDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "genre"
                ]
            },
            "get": {
                "operationId": "GenreController_getAllGenres",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Genre"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "genre"
                ]
            },
            "put": {
                "operationId": "GenreController_updateGenre",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGenreDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "genre"
                ],
                "security": [
                    {
                        "defaultBearerAuth": []
                    }
                ]
            }
        },
        "/genre/{id}": {
            "get": {
                "operationId": "GenreController_getOneGenre",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "genre"
                ]
            },
            "delete": {
                "operationId": "GenreController_deleteGenre",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "genre"
                ]
            }
        }
    },
    "info": {
        "title": "Учебный проект",
        "description": "Клон kinopoisk-a",
        "version": "0.0.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "defaultBearerAuth": {
                "scheme": "bearer",
                "bearerFormat": "jwt",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateUserProfileDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@mail.ru",
                        "description": "Почтовый адрес"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 4,
                        "maxLength": 16,
                        "example": "12345678",
                        "description": "Пароль пользователя"
                    },
                    "name": {
                        "type": "string",
                        "example": "Александр",
                        "description": "Имя пользователя"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Иванов",
                        "description": "Фамилия пользователя"
                    },
                    "nickname": {
                        "type": "string",
                        "example": "stas9n",
                        "description": "Никнейм пользователя"
                    },
                    "city": {
                        "type": "string",
                        "example": "Москва",
                        "description": "Город проживания пользователя"
                    },
                    "url": {
                        "type": "string",
                        "description": "Ссылка на профиль на кинопоиске (имеется только у спаршенных с кинопоиска профилей)"
                    },
                    "photo": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "name",
                    "surname",
                    "nickname",
                    "city"
                ]
            },
...
            "CreateGenreDTO": {
                "type": "object",
                "properties": {
                    "genreName": {
                        "type": "string",
                        "description": "Название жанра",
                        "example": "Комедия"
                    },
                    "genreNameEn": {
                        "type": "string",
                        "description": "Genre title",
                        "example": "Comedy "
                    }
                },
                "required": [
                    "genreName"
                ]
            },
            "Genre": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Первичный ключ",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "description": "Ссылка на жанр на кинопоиске",
                        "example": "https://www.kinopoisk.ru/lists/movies/genre--comedy/?b=films&b=top"
                    },
                    "genreName": {
                        "type": "string",
                        "description": "Название жанра",
                        "example": "Комедия"
                    },
                    "genreNameEn": {
                        "type": "string",
                        "description": "Genre title",
                        "example": "Comedy "
                    }
                },
                "required": [
                    "id",
                    "url",
                    "genreName",
                    "genreNameEn"
                ]
            },
            "UpdateGenreDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Первичный ключ",
                        "example": 1
                    },
                    "genreNameEn": {
                        "type": "string",
                        "description": "Genre title",
                        "example": "Comedy "
                    }
                },
                "required": [
                    "id",
                    "genreNameEn"
                ]
            }
        }
    }
}